+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 * Adds Letter Opener generator. ([@coolprobn][])
 * Adds Whenever generator. ([@aadil][])
+* Adds Rswag generator. ([@coolprobn][])
 
 ## 0.12.0 (May 8th, 2023)
 * Adds Flipper generator. ([@abhaynikam][])